pyproject.toml
[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[project]
name = "zistudy-api"
version = "0.2.0"
description = "Modern asynchronous study and note-taking API."
authors = [
  { name = "Balen Amir" }
]
readme = "README.md"
requires-python = ">=3.14"
license = { text = "MIT" }
keywords = ["study", "fastapi", "notes", "learning"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.14",
  "Framework :: FastAPI",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = [
  "fastapi>=0.120.0",
  "uvicorn[standard]>=0.38.0",
  "sqlalchemy>=2.0.44",
  "asyncpg>=0.30.0",
  "pydantic>=2.12.3",
  "pydantic-settings>=2.11.0",
  "structlog>=25.4.0",
  "alembic>=1.17.0",
  "python-dotenv>=1.1.1",
  "rich>=14.2.0",
  "passlib[argon2]>=1.7.4",
  "pyjwt[crypto]>=2.10.0",
  "email-validator>=2.1.0",
  "httpx>=0.28.1",
  "pymupdf>=1.26.5",
  "python-multipart>=0.0.18",
  "celery[redis]>=5.4.0",
]

[dependency-groups]
dev = [
  "pytest>=8.4.2",
  "pytest-asyncio>=1.2.0",
  "pytest-cov>=7.0.0",
  "httpx>=0.28.1,<1.0",
  "asgi-lifespan>=2.1.0",
  "ruff>=0.14.2",
  "mypy>=1.18.2",
  "anyio>=4.11.0",
  "aiosqlite>=0.19.0",
  "greenlet>=3.0.3",
]

[project.urls]
Homepage = "https://example.com/zistudy"
Repository = "https://example.com/zistudy/repo"

[project.scripts]
zistudy-format = "zistudy_api.tools:run_format"
zistudy-lint = "zistudy_api.tools:run_lint"
zistudy-typecheck = "zistudy_api.tools:run_typecheck"
zistudy-test = "zistudy_api.tools:run_test"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["zistudy_api", "zistudy_api.*"]

[tool.typer]
# Reserved for future Typer configuration.

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "--strict-markers --disable-warnings --maxfail=1"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
src = ["src", "tests"]
target-version = "py314"

[tool.ruff.lint]
extend-select = ["B", "C4", "SIM", "I"]
ignore = ["B905"]

[tool.uv]
prerelease = "allow"

[tool.coverage.run]
branch = true
source = ["src/zistudy_api"]

[tool.coverage.report]
show_missing = true
skip_covered = true

[tool.mypy]
python_version = "3.14"
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_configs = true
show_error_codes = true
strict_optional = true
plugins = ["pydantic.mypy"]
ignore_missing_imports = false
mypy_path = "src"

[[tool.mypy.overrides]]
module = [
  "asyncpg",
  "structlog",
  "structlog.*",
  "uvicorn",
  "passlib",
  "passlib.*",
  "jwt",
  "httpx",
  "httpx.*",
  "celery",
  "celery.*",
]
ignore_missing_imports = true
